fix: 메인 페이지 화면 레이아웃 수정 #24
